+++ airsim_settings/settings.json
