+++ airsim_settings/settings.stereo.cv.json
@@ -1,0 +1,188 @@
+
+{
+    "SettingsVersion": 1.2,
+    "SimMode": "",
+    "ViewMode": "",
+    "Vehicles": {
+      "PhysXCar": {
+        "VehicleType": "PhysXCar",
+        "DefaultVehicleState": "",
+        "AutoCreate": true,
+        "PawnPath": "",
+        "EnableCollisionPassthrogh": false,
+        "EnableCollisions": true,
+        "RC": {
+          "RemoteControlID": -1
+        },
+        "Cameras": {   
+          "0": {
+              "CaptureSettings": [
+                  {
+                      "ImageType": 0,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  },
+                  {
+                      "ImageType": 4,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  }
+                  ],
+                  "X": 0.25,
+                  "Y": -0.1000,
+                  "Z": -1.7,
+                  "Pitch": 0.0,
+                  "Roll": 0.0,
+                  "Yaw": 0
+          },
+  
+          "1": {
+              "CaptureSettings": [
+                  {
+                      "ImageType": 0,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  },
+                  {
+                      "ImageType": 4,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  }
+                  ],
+                  "X": 0.25,
+                  "Y": -0.0500,
+                  "Z": -1.7,
+                  "Pitch": 0.0,
+                  "Roll": 0.0,
+                  "Yaw": 0
+          },
+  
+          "2": {
+              "CaptureSettings": [
+                  {
+                      "ImageType": 0,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  },
+                  {
+                      "ImageType": 4,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  }
+                  ],
+                  "X": 0.25,
+                  "Y": 0.0000,
+                  "Z": -1.7,
+                  "Pitch": 0.0,
+                  "Roll": 0.0,
+                  "Yaw": 0
+          },
+  
+          "3": {
+              "CaptureSettings": [
+                  {
+                      "ImageType": 0,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  },
+                  {
+                      "ImageType": 4,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  }
+                  ],
+                  "X": 0.25,
+                  "Y": 0.0500,
+                  "Z": -1.7,
+                  "Pitch": 0.0,
+                  "Roll": 0.0,
+                  "Yaw": 0
+          },
+  
+          "4": {
+              "CaptureSettings": [
+                  {
+                      "ImageType": 0,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  },
+                  {
+                      "ImageType": 4,
+                      "Width": 480,
+                        "Height": 270,
+                      "FOV_Degrees": 90
+                  }
+                  ],
+                  "X": 0.25,
+                  "Y": 0.1000,
+                  "Z": -1.7,
+                  "Pitch": 0.0,
+                  "Roll": 0.0,
+                  "Yaw": 0
+          }
+  
+          }
+      }
+    },
+    "SubWindows": [
+      {"WindowID": 0, "ImageType": 5, "CameraName": "0", "Visible": true},
+      {"WindowID": 1, "ImageType": 7, "CameraName": "0", "Visible": true}
+    ],
+    "Recording": {
+      "RecordOnMove": false,
+      "RecordInterval": 3,
+      "Cameras": [
+          { "CameraName": "0", "ImageType": 0, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "1", "ImageType": 0, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "2", "ImageType": 0, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "3", "ImageType": 0, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "4", "ImageType": 0, "PixelsAsFloat": false, "Compress": false },
+  
+          { "CameraName": "0", "ImageType": 1, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "1", "ImageType": 1, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "2", "ImageType": 1, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "3", "ImageType": 1, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "4", "ImageType": 1, "PixelsAsFloat": true, "Compress": false },
+  
+          { "CameraName": "0", "ImageType": 2, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "1", "ImageType": 2, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "2", "ImageType": 2, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "3", "ImageType": 2, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "4", "ImageType": 2, "PixelsAsFloat": true, "Compress": false },
+  
+          { "CameraName": "0", "ImageType": 4, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "1", "ImageType": 4, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "2", "ImageType": 4, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "3", "ImageType": 4, "PixelsAsFloat": true, "Compress": false },
+          { "CameraName": "4", "ImageType": 4, "PixelsAsFloat": true, "Compress": false },
+  
+          { "CameraName": "0", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "1", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "2", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "3", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "4", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },		
+  
+          { "CameraName": "0", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "1", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "2", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "3", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "4", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+  
+          { "CameraName": "0", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "1", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "2", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "3", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+          { "CameraName": "4", "ImageType": 7, "PixelsAsFloat": false, "Compress": false }
+      ]
+    }
+  }
+
