+++ airsim_settings/settings.stereo.json
@@ -134,7 +134,8 @@
     }
   },
   "SubWindows": [
-	{"WindowID": 0, "ImageType": 0, "CameraName": "0", "Visible": true}
+	{"WindowID": 0, "ImageType": 5, "CameraName": "0", "Visible": true},
+	{"WindowID": 1, "ImageType": 7, "CameraName": "0", "Visible": true}
   ],
   "Recording": {
     "RecordOnMove": false,
@@ -170,11 +171,17 @@
 		{ "CameraName": "3", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },
 		{ "CameraName": "4", "ImageType": 5, "PixelsAsFloat": false, "Compress": false },		
 
-		{ "CameraName": "0", "ImageType": 7, "PixelsAsFloat": true, "Compress": false },
-		{ "CameraName": "1", "ImageType": 7, "PixelsAsFloat": true, "Compress": false },
-		{ "CameraName": "2", "ImageType": 7, "PixelsAsFloat": true, "Compress": false },
-		{ "CameraName": "3", "ImageType": 7, "PixelsAsFloat": true, "Compress": false },
-		{ "CameraName": "4", "ImageType": 7, "PixelsAsFloat": true, "Compress": false }
+		{ "CameraName": "0", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "1", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "2", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "3", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "4", "ImageType": 6, "PixelsAsFloat": false, "Compress": false },
+
+		{ "CameraName": "0", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "1", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "2", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "3", "ImageType": 7, "PixelsAsFloat": false, "Compress": false },
+		{ "CameraName": "4", "ImageType": 7, "PixelsAsFloat": false, "Compress": false }
     ]
   }
 }
