+++ airsim_settings/settings.stereo.small.json
