+++ airsim_utils/generate_cameras.py
