+++ airsim_utils/meshes/meshes_analysis.txt
